requirements.txt
playwright
python-telegram-bot